requirements.txt
amqp==5.0.9
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.4.0
boto3==1.20.37
botocore==1.23.37
cached-property==1.5.2
celery==5.1.2
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.2.11
django-filter==21.1
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.13.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.8.3
incremental==16.10.1
jmespath==0.10.0
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.1.0
language-selector==0.1
mysqlclient==2.0.3
netifaces==0.10.4
PAM==0.4.2
prompt-toolkit==3.0.24
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.2
python-debian==0.1.32
python-memcached==1.59
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.5.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.7
swapper==1.3.0
systemd-python==234
Twisted==17.9.0
typing-extensions==4.0.1
ufw==0.36
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
zope.interface==4.3.2
